requirements-dev.txt
# requirements-dev.txt
black==23.9.1
pytest==7.4.2
pytest-django==4.5.2